+++ src/components/base/Base.js
@@ -6,6 +6,7 @@ import Tooltip from 'tooltip.js';
 import i18next from 'i18next';
 import FormioUtils from '../../utils';
 import {Validator} from '../Validator';
+import moment from 'moment';
 
 /**
  * This is the BaseComponent class which all elements within the FormioForm derive from.
@@ -1866,7 +1867,6 @@ export class BaseComponent {
     return BaseComponent.requireLibrary('quill', 'Quill', 'https://cdn.quilljs.com/1.3.6/quill.min.js', true)
       .then(() => {
         this.quill = new Quill(element, settings);
-        this.quill.root.spellcheck = this.component.spellcheck;
 
         /** This block of code adds the [source] capabilities.  See https://codepen.io/anon/pen/ZyEjrQ **/
         const txtArea = document.createElement('textarea');
@@ -2110,7 +2110,8 @@ export class BaseComponent {
       component: this.component,
       data,
       row: this.data,
-      instance: this
+      instance: this,
+      moment
     }, 'value'), flags);
   }