+++ src/components/base/editForm/Base.edit.conditional.js
@@ -1,3 +1,5 @@
+import { EditFormUtils } from "./utils";
+
 export let BaseEditConditional = [
   {
     type: 'panel',
@@ -45,60 +47,10 @@ export let BaseEditConditional = [
       }
     ]
   },
-  {
-    type: 'panel',
-    title: 'Advanced',
-    key: 'advanced-conditional',
-    theme: 'default',
-    collapsible: true,
-    collapsed: true,
-    components: [
-      {
-        type: 'textarea',
-        key: 'customConditional',
-        rows: 5,
-        editor: 'ace',
-        input: true,
-        placeholder: `show = (data['mykey'] > 1);`
-      },
-      {
-        type: 'htmlelement',
-        tag: 'div',
-        content: '<small>' +
-          '<p>Enter custom conditional code.</p>' +
-          '<p>You must assign the <strong>show</strong> variable as either <strong>true</strong> or <strong>false</strong>.</p>' +
-          '<p>The global variable <strong>data</strong> is provided, and allows you to access the data of any form component, by using its API key.</p>' +
-          '<p><strong>Note: Advanced Conditional logic will override the results of the Simple Conditional logic.</strong></p>' +
-        '</small>'
-      }
-    ]
-  },
-  {
-    type: 'panel',
-    title: 'JSON Conditional',
-    key: 'json-conditional',
-    theme: 'default',
-    collapsible: true,
-    collapsed: true,
-    components: [
-      {
-        type: 'textarea',
-        key: 'conditional.json',
-        rows: 5,
-        editor: 'ace',
-        as: 'json',
-        input: true,
-        placeholder: `{ ... }`
-      },
-      {
-        type: 'htmlelement',
-        tag: 'div',
-        content: '<small>' +
-          '<p>Execute custom validation logic with JSON and <a href="http://jsonlogic.com/">JsonLogic</a>.</p>' +
-          '<p>Submission data is available as JsonLogic variables, with the same api key as your components.</p>' +
-          '<p><a href="http://formio.github.io/formio.js/app/examples/conditions.html" target="_blank">Click here for an example</a></p>' +
-        '</small>'
-      }
-    ]
-  }
+  EditFormUtils.javaScriptValue('Advanced Conditions', 'customConditional', 'conditional.json', 110,
+    '<p>You must assign the <strong>show</strong> variable as either <strong>true</strong> or <strong>false</strong>.</p>' +
+    '<p><strong>Note: Advanced Conditional logic will override the results of the Simple Conditional logic.</strong></p>' +
+    '<h5>Example</h5><pre>show = !!data.showMe;</pre>',
+    '<p><a href="http://formio.github.io/formio.js/app/examples/conditions.html" target="_blank">Click here for an example</a></p>'
+  )
 ];