+++ src/components/checkbox/Checkbox.js
@@ -128,7 +128,7 @@ export class CheckBoxComponent extends BaseComponent {
   }
 
   createLabel(container, input) {
-    if (super.labelIsHidden()) {
+    if (!this.component.label) {
       return null;
     }
 
@@ -153,7 +153,7 @@ export class CheckBoxComponent extends BaseComponent {
     if (this.info.attr.id) {
       this.labelElement.setAttribute('for', this.info.attr.id);
     }
-    if (!this.options.inputsOnly && labelOnTheTopOrOnTheLeft) {
+    if (!this.labelIsHidden() && labelOnTheTopOrOnTheLeft) {
       this.setInputLabelStyle(this.labelElement);
       this.setInputStyle(input);
       this.labelSpan.appendChild(this.text(this.component.label));
@@ -161,7 +161,7 @@ export class CheckBoxComponent extends BaseComponent {
     }
     this.addInput(input, this.labelElement);
 
-    if (!this.options.inputsOnly && !labelOnTheTopOrOnTheLeft) {
+    if (!this.labelIsHidden() && !labelOnTheTopOrOnTheLeft) {
       this.setInputLabelStyle(this.labelElement);
       this.setInputStyle(input);
       this.labelSpan.appendChild(this.text(this.addShortcutToLabel()));
@@ -178,30 +178,6 @@ export class CheckBoxComponent extends BaseComponent {
     const input = this.ce(this.info.type, this.info.attr);
     this.errorContainer = container;
     return input;
-  }
-
-  updateValueByName() {
-    const component = this.getRoot().getComponent(this.component.name);
-    if (component) {
-      component.setValue(this.component.value, {changed: true});
-    }
-    else {
-      _.set(this.data, this.component.name, this.component.value);
-    }
-  }
-
-  addInputEventListener(input) {
-    this.addEventListener(input, this.info.changeEvent, () => {
-      // If this input has a "name", then its other input elements are elsewhere on
-      // the form. To get the correct submission object, we need to refresh the whole
-      // data object.
-      if (this.component.name) {
-        this.updateValueByName();
-        this.emit('refreshData');
-      }
-
-      this.updateValue();
-    });
   }
 
   getValueAt(index) {
@@ -234,7 +210,7 @@ export class CheckBoxComponent extends BaseComponent {
 
   get dataValue() {
     if (this.component.name) {
-      return _.get(this.data, this.component.name, this.emptyValue);
+      return _.get(this.data, this.component.name, this.emptyValue) === this.component.value;
     }
 
     return super.dataValue;
@@ -242,7 +218,9 @@ export class CheckBoxComponent extends BaseComponent {
 
   set dataValue(value) {
     if (this.component.name) {
-      _.set(this.data, this.component.name, value);
+      if (value) {
+        _.set(this.data, this.component.name, this.component.value);
+      }
       return value;
     }